Mudado para Long o tipo de retorno de Evento.getChaveLocal()

+++ app/src/main/java/com/infact/nightour/model/Evento.java
@@ -128,7 +128,7 @@ public class Evento {
         return getImagem().getId();
     }
 
-    public int getChaveLocal() {
+    public Long getChaveLocal() {
         return getLocal().getId();
     }
